requirements.txt
sarvamai==0.1.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
gTTS==2.4.0
pygame==2.5.2
httpx==0.27.0
simpleaudio==1.0.4 